Remove ems functions, moving parts to own file
